RF Combining ManageJenkins class
